# pipeline-manager: support multi-crate compilation