The compiler server now passes the `--crates` argument to SQL compiler, causing it to generate multiple crates instead of a single `main.rs`.

This yields two main benefits:
- Partial (re)compilation: compile only the portions of the circuit which are affected and have not yet been compiled
- Faster compilation: the decomposition into multiple crates enables more parallelism in the cargo build process

Notable changes to the compiler server inner workings:
- The `main_rust` field now contains the base64-encoded gzipped archive of the directory of crates that the SQL compiler outputs
- Cleanup now also includes the `target/<target>/incremental` directory, which contains artifacts when building with the `dev` profile for which `<target>` is `debug`
- Source checksum no longer covers compiler argument paths, as they generally change only if the platform version changes, which is covered
- Compilation artifacts are now also kept of pipelines which are not (yet) successfully compiled, which prevents them from being cleaned up if there is a large Rust (re)compilation queue
- Compilation artifact retention is increased from 900s to 3600s, and the cleanup interval from 30s to 120s
- In terms of `rust-compilation` directory content:
  - Directories `projects` and `binaries` are no longer used and can be manually deleted
  - Directory `crates` is added, and is the directory where all crates used for compilation are kept
  - Directory `pipeline-binaries` has not changed, as such binaries from Rust compilations prior to this commit are still usable
  - Pipeline-related artifacts in `target` from Rust compilations prior to this commit are no longer detected during cleanup due to the different artifact naming. These artifacts can be manually deleted.
  - Cleanup state is now stored in `target_cleanup.json`. The prior cleanup state file `cleanup_state.json` can be manually deleted.