@snkas 2025-03-25
Thanks for the review!

The reasoning is that the SQL compiler and Rust compiler are separate, and communicate through the database. The benefit is (1) there is no need for synchronization between the two, where the SQL compiler cannot write while the Rust compiler is reading for example, and (2) allows potentially scaling up the number of Rust compiler servers independent of SQL compiler servers.

The current type of the `main_rust` is a String, changing it to Bytes would be backward incompatible, hence the base64 encoding.