@gz 2025-03-25
>The current type of the main_rust is a String, changing it to Bytes would be backward incompatible, hence the base64 encoding.

I am not sure, it should be ok to make the field bytes, e.g., a string should be able to go into something that is now of type `bytes`.