@snkas 2025-03-27
Compilation artifact retention is increased from 900s to 3600s, and the cleanup interval from 30s to 120s